adds channel default characters

+++ src/commands/chat/characters/character-set-default.subcommand.ts
@@ -1,5 +1,9 @@
 import { WanderersGuide } from '../../../services/wanderers-guide/index';
-import { Character, GuildDefaultCharacter } from '../../../services/kobold/models/index.js';
+import {
+	ChannelDefaultCharacter,
+	Character,
+	GuildDefaultCharacter,
+} from '../../../services/kobold/models/index.js';
 import {
 	ApplicationCommandType,
 	RESTPostAPIChatInputApplicationCommandsJSONBody,
@@ -17,13 +21,14 @@ import { InteractionUtils } from '../../../utils/index.js';
 import { Command, CommandDeferType } from '../../index.js';
 import { Language } from '../../../models/enum-helpers/index.js';
 import { TranslationFunctions } from '../../../i18n/i18n-types.js';
+import { CharacterOptions } from './command-options.js';
 
-export class CharacterSetServerDefaultSubCommand implements Command {
-	public names = [Language.LL.commands.character.setServerDefault.name()];
+export class CharacterSetDefaultSubCommand implements Command {
+	public names = [Language.LL.commands.character.setDefault.name()];
 	public metadata: RESTPostAPIChatInputApplicationCommandsJSONBody = {
 		type: ApplicationCommandType.ChatInput,
-		name: Language.LL.commands.character.setServerDefault.name(),
-		description: Language.LL.commands.character.setServerDefault.description(),
+		name: Language.LL.commands.character.setDefault.name(),
+		description: Language.LL.commands.character.setDefault.description(),
 		dm_permission: true,
 		default_member_permissions: undefined,
 	};
@@ -48,10 +53,10 @@ export class CharacterSetServerDefaultSubCommand implements Command {
 
 			if (
 				match == '' ||
-				Language.LL.commands.character.setServerDefault.noneOption().includes(match)
+				Language.LL.commands.character.setDefault.noneOption().includes(match)
 			) {
 				options.push({
-					name: Language.LL.commands.character.setServerDefault.noneOption().toString(),
+					name: Language.LL.commands.character.setDefault.noneOption().toString(),
 					value: '__NONE__',
 				});
 			}
@@ -66,8 +71,12 @@ export class CharacterSetServerDefaultSubCommand implements Command {
 		data: EventData,
 		LL: TranslationFunctions
 	): Promise<void> {
+		const defaultScope = intr.options.getString(
+			CharacterOptions.CHARACTER_SET_DEFAULT_SCOPE.name
+		);
 		const charName = intr.options.getString(ChatArgs.SET_ACTIVE_NAME_OPTION.name);
 		const currentGuildId = intr.guildId;
+		const currentChannelId = intr.channelId;
 
 		// try and find that charcter
 		const targetCharacter = (
@@ -75,39 +84,71 @@ export class CharacterSetServerDefaultSubCommand implements Command {
 		)[0];
 
 		if (targetCharacter) {
-			//set all other characters as not the default for this user in this guild
-			await GuildDefaultCharacter.query()
-				.delete()
-				.where({ userId: intr.user.id, guildId: currentGuildId });
+			if (
+				defaultScope === Language.LL.commandOptions.setDefaultScope.choices.channel.value()
+			) {
+				//set all other characters as not the default for this user in this channel
+				await ChannelDefaultCharacter.query()
+					.delete()
+					.where({ userId: intr.user.id, channelId: currentChannelId });
 
-			//set the character as the default for this guild
-			await GuildDefaultCharacter.query().insert({
-				userId: intr.user.id,
-				guildId: currentGuildId,
-				characterId: targetCharacter.id,
-			});
+				//set the character as the default for this channel
+				await ChannelDefaultCharacter.query().insert({
+					userId: intr.user.id,
+					channelId: currentChannelId,
+					characterId: targetCharacter.id,
+				});
+			} else if (
+				defaultScope === Language.LL.commandOptions.setDefaultScope.choices.server.value()
+			) {
+				//set all other characters as not the default for this user in this guild
+				await GuildDefaultCharacter.query()
+					.delete()
+					.where({ userId: intr.user.id, guildId: currentGuildId });
 
+				//set the character as the default for this guild
+				await GuildDefaultCharacter.query().insert({
+					userId: intr.user.id,
+					guildId: currentGuildId,
+					characterId: targetCharacter.id,
+				});
+			}
 			//send success message
 			await InteractionUtils.send(
 				intr,
-				LL.commands.character.setServerDefault.interactions.success({
+				LL.commands.character.setDefault.interactions.success({
 					characterName: targetCharacter.name,
+					scope: defaultScope,
 				})
 			);
 		} else {
 			if ('__NONE__'.includes(charName)) {
 				//unset the server default character.
-				await GuildDefaultCharacter.query()
-					.delete()
-					.where({ userId: intr.user.id, guildId: currentGuildId });
+				if (
+					defaultScope ===
+					Language.LL.commandOptions.setDefaultScope.choices.channel.value()
+				) {
+					await ChannelDefaultCharacter.query()
+						.delete()
+						.where({ userId: intr.user.id, channelId: currentChannelId });
+				} else if (
+					defaultScope ===
+					Language.LL.commandOptions.setDefaultScope.choices.server.value()
+				) {
+					await GuildDefaultCharacter.query()
+						.delete()
+						.where({ userId: intr.user.id, guildId: currentGuildId });
+				}
 				await InteractionUtils.send(
 					intr,
-					LL.commands.character.setServerDefault.interactions.removed()
+					LL.commands.character.setDefault.interactions.removed({
+						scope: defaultScope,
+					})
 				);
 			} else {
 				await InteractionUtils.send(
 					intr,
-					LL.commands.character.setServerDefault.interactions.notFound()
+					LL.commands.character.setDefault.interactions.notFound()
 				);
 			}
 		}
